Revert "Daily account snapshot"
