+++ database/cosmos_extensions.go
@@ -13,18 +13,19 @@ func (db BigDipperDb) SaveValidators(validators []staking.Validator) error {
 
 	queryInsert := "INSERT INTO validator (consensus_address, consensus_pubkey) VALUES "
 	for i, result := range validators {
-		p1 := i * 2 // Starting position for insert params
+		p1 := i * 2 // starting position for insert params
 
-		publicKey, err := sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeConsPub, result.ConsPubKey)
+		queryInsert += fmt.Sprintf("($%d,$%d),", p1+1, p1+2)
+
+		key, err := sdk.Bech32ifyPubKey(sdk.Bech32PubKeyTypeConsPub, result.ConsPubKey)
 		if err != nil {
 			return err
 		}
 
-		queryInsert += fmt.Sprintf("($%d,$%d),", p1+1, p1+2)
-		insertParams = append(insertParams, result.ConsAddress().String(), publicKey)
+		insertParams = append(insertParams, result.ConsAddress().String(), key)
 	}
 
-	queryInsert = queryInsert[:len(queryInsert)-1] // Remove trailing ","
+	queryInsert = queryInsert[:len(queryInsert)-1] // remove trailing ","
 	queryInsert += " ON CONFLICT DO NOTHING"
 	_, err := db.Sql.Exec(queryInsert, insertParams...)
 	return err
