+++ go.mod
@@ -7,7 +7,5 @@ require (
 	github.com/desmos-labs/juno v0.0.0-20200615084022-e95f19b114df
 	github.com/go-co-op/gocron v0.2.0
 	github.com/jmoiron/sqlx v1.2.1-0.20200324155115-ee514944af4b
-	github.com/lib/pq v1.3.0
 	github.com/rs/zerolog v1.18.0
-	github.com/tendermint/tendermint v0.33.3
 )
